Fixed DFS comparaison error

+++ include/DFS.hpp
@@ -99,7 +99,7 @@ class DFS : public Solver<GraphMove, GraphComm> {
   bool IsConfigurationConnected(const std::shared_ptr<Configuration>& config) {
     std::vector<bool> agent_treated = std::vector<bool>(this->instance_.nb_agents(), false);
     std::stack<Agent> agent_stack;
-    int agent_count = 0;
+    size_t agent_count = 0;
     std::vector<bool>::iterator it;
 
     while ((it = find(agent_treated.begin(), agent_treated.end(), false)) != agent_treated.end()) {
